dump pir py code before execute
